requirements.txt
altair==5.5.0
altgraph==0.17.2
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.5.2
astroid==2.4.2
asttokens==2.4.1
async-lru==2.0.4
asyncgui==0.6.3
asynckivy==0.6.3
attrs==23.2.0
Babel==2.15.0
bcrypt==4.3.0
beautifulsoup4==4.11.1
beeware==0.3.0
binaryornot==0.4.4
bleach==6.1.0
blinker==1.9.0
briefcase==0.3.19
build==1.2.1
cachetools==5.5.2
captcha==0.7.1
certifi==2021.10.8
cffi==1.16.0
chardet==5.2.0
charset-normalizer==2.0.9
click==8.1.7
clr-loader==0.2.6
colorama==0.4.4
comm==0.2.2
contourpy==1.2.1
cookiecutter==2.6.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Django==4.0.6
docutils==0.18.1
et_xmlfile==2.0.0
exceptiongroup==1.2.1
executing==2.0.1
extra-streamlit-components==0.1.80
fastjsonschema==2.19.1
fonttools==4.51.0
fqdn==1.5.1
future==0.18.2
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.40.3
google-auth-oauthlib==1.2.2
greenlet==3.0.3
gspread==6.2.1
gurobipy==12.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.3
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.18.1
ipywidgets==8.1.2
isoduration==20.11.0
isort==5.7.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.4
kivymd @ https://github.com/kivymd/KivyMD/archive/master.zip#sha256=16aac37c1f54ea0ba6970d478ee9fba286ff916fd0f30186cd96b44e577eb479
kiwisolver==1.4.5
lazy-object-proxy==1.4.3
lxml==4.9.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
materialyoucolor==2.0.9
matplotlib==3.8.4
matplotlib-inline==0.1.7
mccabe==0.6.1
mdurl==0.1.2
mistune==3.0.2
narwhals==1.43.1
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
notebook==7.1.3
notebook_shim==0.2.4
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.3.0
openpyxl==3.1.5
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
patsy==0.5.6
pefile==2021.9.3
pillow==10.3.0
platformdirs==4.2.1
playsound==1.2.2
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==6.31.1
psutil==5.9.8
PuLP==2.9.0
pure-eval==0.2.2
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.3
PyJWT==2.10.1
pylint==2.6.0
pyparsing==3.1.2
pypiwin32==223
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-json-logger==2.0.7
python-slugify==8.0.4
pythonnet==3.0.3
pytz==2024.1
pywin32==303
pywin32-ctypes==0.2.0
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.0
rpds-py==0.18.1
rsa==4.9.1
salabim==24.0.11.post1
scikit-learn==1.4.2
scipy==1.13.0
Send2Trash==1.8.3
six==1.15.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.6.3
statsmodels==0.14.2
streamlit==1.46.0
streamlit-authenticator==0.4.2
streamlit-js-eval==0.1.7
tenacity==9.1.2
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.5.0
tinycss2==1.3.0
toga==0.4.6
toga-core==0.4.6
toga-winforms==0.4.6
toml==0.10.2
tomli==2.0.1
tomli_w==1.0.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
travertino==0.3.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
uri-template==1.3.0
urllib3==1.26.7
watchdog==6.0.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.10
wrapt==1.12.1
xlrd==2.0.1
zipp==3.18.1
